assortment-planner-agent/requirements.txt
google-adk
fastapi
uvicorn[standard]
litellm
langfuse
python-dotenv
pydantic
httpx

